+++ test/all.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'package:io/ansi.dart';
+import '404_hole_test.dart' as hole404;
 import 'accepts_test.dart' as accepts;
 import 'anonymous_service_test.dart' as anonymous_service;
 import 'body_test.dart' as body;
@@ -30,6 +31,7 @@ import 'package:test/test.dart';
 /// For running with coverage
 main() {
   print(cyan.wrap('Running tests on ${Platform.version}'));
+  group('404_hole', hole404.main);
   group('accepts', accepts.main);
   group('anonymous service', anonymous_service.main);
   group('body', body.main);